feat: renamed accountbound v2 in accountbound
