+++ contracts/attesters/hydra-s1/interfaces/IHydraS1AccountboundAttester.sol
@@ -3,41 +3,14 @@
 pragma solidity ^0.8.17;
 pragma experimental ABIEncoderV2;
 
-import {Attestation} from '../../../core/libs/Structs.sol';
-import {IAttester} from '../../../core/interfaces/IAttester.sol';
-import {CommitmentMapperRegistry} from '../../../periphery/utils/CommitmentMapperRegistry.sol';
-import {AvailableRootsRegistry} from '../../../periphery/utils/AvailableRootsRegistry.sol';
-import {HydraS1Lib, HydraS1ProofData, HydraS1ProofInput} from './../libs/HydraS1Lib.sol';
-import {IHydraS1Base} from './../base/IHydraS1Base.sol';
+import {IHydraS1SimpleAttester} from '././IHydraS1SimpleAttester.sol';
 
 /**
  * @title Hydra-S1 Accountbound Interface
  * @author Sismo
- * @notice Interface with the errors, events and methods specific to the HydraS1AcountboundAttester.
+ * @notice Interface of the HydraS1AccountboundAttester contract which inherits from the errors, events and methods specific to the HydraS1SimpleAttester interface.
  **/
-interface IHydraS1AccountboundAttester is IHydraS1Base, IAttester {
-  struct NullifierData {
-    address destination;
-    uint32 cooldownStart;
-    uint16 burnCount;
-  }
-
-  /**
-   * @dev Error when the nullifier is on cooldown. The user have to wait the cooldownDuration
-   * before being able to change again the destination address.
-   **/
-  error NullifierOnCooldown(NullifierData nullifierData, uint32 cooldownDuration);
-
-  /**
-   * @dev Error when the collectionId of an attestation overflow the AUTHORIZED_COLLECTION_ID_LAST
-   **/
-  error CollectionIdOutOfBound(uint256 collectionId);
-
-  /**
-   * @dev Event emitted when the nullifier (or nullifierHash) is associated to a destination address.
-   **/
-  event NullifierDestinationUpdated(uint256 nullifier, address newOwner);
-
+interface IHydraS1AccountboundAttester is IHydraS1SimpleAttester {
   /**
    * @dev Event emitted when the nullifier has been set on cooldown. This happens when the
    * attestation destination of a nullifier has been changed
@@ -45,26 +18,44 @@ interface IHydraS1AccountboundAttester is IHydraS1Base, IAttester {
   event NullifierSetOnCooldown(uint256 nullifier, uint16 burnCount);
 
   /**
-   * @dev Getter, returns the data linked to a nullifier
+   * @dev Error when the nullifier is on cooldown. The user have to wait the cooldownDuration
+   * before being able to change again the destination address.
+   **/
+  error NullifierOnCooldown(
+    uint256 nullifier,
+    address destination,
+    uint16 burnCount,
+    uint32 cooldownStart
+  );
+
+  /**
+   * @dev Error when the cooldown duration for a given groupId is equal to zero.
+   * The HydraS1AccountboundAttester behaves like the HydraS1SimpleAttester.
+   **/
+  error CooldownDurationNotSetForGroupId(uint256 groupId);
+
+  /**
+   * @dev Getter, returns the cooldown start of a nullifier
    * @param nullifier nullifier used
    **/
-  function getNullifierData(uint256 nullifier) external view returns (NullifierData memory);
+  function getNullifierCooldownStart(uint256 nullifier) external view returns (uint32);
 
   /**
-   * @dev Getter, returns the last attestation destination of a nullifier
+   * @dev Getter, returns the burnCount of a nullifier
    * @param nullifier nullifier used
    **/
-  function getDestinationOfNullifier(uint256 nullifier) external view returns (address);
+  function getNullifierBurnCount(uint256 nullifier) external view returns (uint16);
 
   /**
-   * @dev Getter
-   * returns of the first collection in which the attester is supposed to record
+   * @dev Setter, sets the cooldown duration of a groupId
+   * @param groupId Id of the group
+   * @param cooldownDuration cooldown duration we want to set for the groupId
    **/
-  function AUTHORIZED_COLLECTION_ID_FIRST() external view returns (uint256);
+  function setCooldownDurationForgroupId(uint256 groupId, uint32 cooldownDuration) external;
 
   /**
-   * @dev Getter
-   * returns of the last collection in which the attester is supposed to record
+   * @dev Getter, get the cooldown duration of a groupId
+   * @param groupId Id of the group
    **/
-  function AUTHORIZED_COLLECTION_ID_LAST() external view returns (uint256);
+  function getCooldownDurationForGroupId(uint256 groupId) external view returns (uint32);
 }
