+++ tasks/deploy-tasks/unit/attesters/hydra-s1/variants/deploy-hydra-s1-accountbound-attester.task.ts
@@ -96,6 +96,7 @@ async function deploymentAction(
     deployed.address,
     deployer
   );
+
   return { hydraS1AccountboundAttester, hydraS1Verifier };
 }
